refactor: putting the dateVisualTransformer for the date text field to a separate file and removing useless comments
